# KAFKA-14355: Fix integer overflow in ProducerPerformance