In `kafka-producer-perf-test.sh`, an integer overflow occurs if a fairly large value is passed to `--num-records`. It causes a NegativeArraySizeException error. This PR fixed this problem by changing types.

Before:
```
$ ./bin/kafka-producer-perf-test.sh --topic=test --producer-props bootstrap.servers=localhost:9092 --num-records 2000000000000000 --throughput 1 --record-size 10
Exception in thread "main" java.lang.NegativeArraySizeException
    at org.apache.kafka.tools.ProducerPerformance$Stats.<init>(ProducerPerformance.java:354)
    at org.apache.kafka.tools.ProducerPerformance.start(ProducerPerformance.java:97)
    at org.apache.kafka.tools.ProducerPerformance.main(ProducerPerformance.java:52)
```


After:
```
$ ./bin/kafka-producer-perf-test.sh --topic=test --producer-props bootstrap.servers=localhost:9092 --num-records 2000000000000000 --throughput 1 --record-size 10
7 records sent, 1.3 records/sec (0.00 MB/sec), 123.7 ms avg latency, 684.0 ms max latency.
5 records sent, 1.0 records/sec (0.00 MB/sec), 2.2 ms avg latency, 3.0 ms max latency.
5 records sent, 1.0 records/sec (0.00 MB/sec), 2.2 ms avg latency, 3.0 ms max latency.
5 records sent, 1.0 records/sec (0.00 MB/sec), 2.6 ms avg latency, 3.0 ms max latency.
...
```

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)